ly_requirements.txt
# ===============================
# Core async / networking
# ===============================
asyncpg>=0.29.0,<1.0
aiohttp>=3.9.0,<4.0
python-dotenv>=1.0.0,<2.0

# ===============================
# Telegram
# ===============================
telethon>=1.34.0,<2.0

# ===============================
# NLP / Topic Modeling (Chinese)
# ===============================
bertopic>=0.16.0,<0.17
sentence-transformers>=2.2.2,<3.0
scikit-learn>=1.3.0,<1.6
numpy>=1.24.0,<2.0
scipy>=1.10.0,<1.14

# ===============================
# Keyword extraction (Chinese-friendly)
# ===============================
pke-zh>=0.1.0

# ===============================
# Utilities
# ===============================
tqdm>=4.66.0,<5.0
